// blockchain/typechain-types/SensorDataRegistry.ts
export interface SensorDataRegistryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "getDevice"
      | "getSensorHash"
      | "owner"
      | "recordSensorHash"
      | "registerDevice"
      | "verifySensorHash"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "DeviceRegistered" | "SensorHashRecorded"
  ): EventFragment;

  encodeFunctionData(functionFragment: "getDevice", values: [string]): string;
  encodeFunctionData(
    functionFragment: "getSensorHash",
    values: [string, BigNumberish]
  ): string;
  encodeFunctionData(functionFragment: "owner", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "recordSensorHash",
    values: [string, BigNumberish, BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "registerDevice",
    values: [string, AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "verifySensorHash",
    values: [string, BigNumberish, BytesLike]
  ): string;

  decodeFunctionResult(functionFragment: "getDevice", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "getSensorHash",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "owner", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "recordSensorHash",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "registerDevice",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "verifySensorHash",
    data: BytesLike
  ): Result;
}

export namespace DeviceRegisteredEvent {
  export type InputTuple = [deviceId: string, authority: AddressLike];
  export type OutputTuple = [deviceId: string, authority: string];
  export interface OutputObject {
    deviceId: string;
    authority: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace SensorHashRecordedEvent {
  export type InputTuple = [
    deviceId: string,
    packetNumber: BigNumberish,
    dataHash: BytesLike,
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    deviceId: string,
    packetNumber: bigint,
    dataHash: string,
    timestamp: bigint
  ];
  export interface OutputObject {
    deviceId: string;
    packetNumber: bigint;
    dataHash: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface SensorDataRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): SensorDataRegistry;
  waitForDeployment(): Promise<this>;

  interface: SensorDataRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getDevice: TypedContractMethod<
    [deviceId: string],
    [[string, boolean] & { authority: string; isRegistered: boolean }],
    "view"
  >;

  getSensorHash: TypedContractMethod<
    [deviceId: string, packetNumber: BigNumberish],
    [string],
    "view"
  >;

  owner: TypedContractMethod<[], [string], "view">;

  recordSensorHash: TypedContractMethod<
    [deviceId: string, packetNumber: BigNumberish, dataHash: BytesLike],
    [void],
    "nonpayable"
  >;

  registerDevice: TypedContractMethod<
    [deviceId: string, authority: AddressLike],
    [void],
    "nonpayable"
  >;

  verifySensorHash: TypedContractMethod<
    [deviceId: string, packetNumber: BigNumberish, dataHash: BytesLike],
    [boolean],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getDevice"
  ): TypedContractMethod<
    [deviceId: string],
    [[string, boolean] & { authority: string; isRegistered: boolean }],
    "view"
  >;
  getFunction(
    nameOrSignature: "getSensorHash"
  ): TypedContractMethod<
    [deviceId: string, packetNumber: BigNumberish],
    [string],
    "view"
  >;
  getFunction(
    nameOrSignature: "owner"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "recordSensorHash"
  ): TypedContractMethod<
    [deviceId: string, packetNumber: BigNumberish, dataHash: BytesLike],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "registerDevice"
  ): TypedContractMethod<
    [deviceId: string, authority: AddressLike],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "verifySensorHash"
  ): TypedContractMethod<
    [deviceId: string, packetNumber: BigNumberish, dataHash: BytesLike],
    [boolean],
    "view"
  >;

  getEvent(
    key: "DeviceRegistered"
  ): TypedContractEvent<
    DeviceRegisteredEvent.InputTuple,
    DeviceRegisteredEvent.OutputTuple,
    DeviceRegisteredEvent.OutputObject
  >;
  getEvent(
    key: "SensorHashRecorded"
  ): TypedContractEvent<
    SensorHashRecordedEvent.InputTuple,
    SensorHashRecordedEvent.OutputTuple,
    SensorHashRecordedEvent.OutputObject
  >;

  filters: {
    "DeviceRegistered(string,address)": TypedContractEvent<
      DeviceRegisteredEvent.InputTuple,
      DeviceRegisteredEvent.OutputTuple,
      DeviceRegisteredEvent.OutputObject
    >;
    DeviceRegistered: TypedContractEvent<
      DeviceRegisteredEvent.InputTuple,
      DeviceRegisteredEvent.OutputTuple,
      DeviceRegisteredEvent.OutputObject
    >;

    "SensorHashRecorded(string,uint256,bytes32,uint256)": TypedContractEvent<
      SensorHashRecordedEvent.InputTuple,
      SensorHashRecordedEvent.OutputTuple,
      SensorHashRecordedEvent.OutputObject
    >;
    SensorHashRecorded: TypedContractEvent<
      SensorHashRecordedEvent.InputTuple,
      SensorHashRecordedEvent.OutputTuple,
      SensorHashRecordedEvent.OutputObject
    >;
  };
}